+++ common/Random.h
@@ -13,8 +13,8 @@ private:
 public:
     static u4 uniformU4();
     static u4 uniformU4(u4 from, u4 to);
-    static u8 uniformU8();
-    static u8 uniformU8(u8 from, u8 to);
+    static uint64_t uniformU8();
+    static uint64_t uniformU8(uint64_t from, uint64_t to);
 };
 } // namespace sorbet
